convert prepare-permutive to ts
